JUPYTER NOTEBOOKS WITH PYTHON
 loans' applications in a commercial bank.

The dataset is to be used to show you how to find the solutions below;

What is the marginal probability of any continuous variable? 
What is the marginal probability of any categorical variable? 
What is the probability that a "female" applicant has her Loan_Status reflecting "Y?" 
Explain the central tendency of any two continuous variables.
Explain the relationship between the variables "Gender" and "Loan_Status."